+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Changed
 
 - Drop support for PHP < 8.1
+- Rename Stack constructor parameter `$defaultResponse` to `$response`
 
 ## [2.0.1] - 2021-03-14